# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CampaignCell API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/cases:
    get:
      operationId: listCases
      tags: [cases]
      summary: List optimization cases
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of optimization cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseListResponse"
  /v1/cases/{id}:
    get:
      operationId: getCase
      tags: [cases]
      summary: Get optimization case detail
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizationCase"
    patch:
      operationId: updateCase
      tags: [cases]
      summary: Update optimization case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseUpdate"
      responses:
        "200":
          description: Case updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizationCase"
  /v1/cases/{id}/assign:
    post:
      operationId: assignCase
      tags: [cases]
      summary: Assign case to an expert
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseAssignInput"
      responses:
        "200":
          description: Case assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizationCase"
  /v1/cases/{id}/complete:
    post:
      operationId: completeCase
      tags: [cases]
      summary: Mark case as completed
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseCompleteInput"
      responses:
        "200":
          description: Case completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OptimizationCase"
  /v1/cases/{id}/notes:
    post:
      operationId: addCaseNote
      tags: [cases]
      summary: Add optimization note to case
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseNoteInput"
      responses:
        "201":
          description: Note added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseNote"

  # ─── SUBSCRIBERS ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OptimizationCase:
      type: object
      required: [id, caseCode, campaignId, status, priority, segment, slaDeadline, createdAt]
      properties:
        id:
          type: string
        caseCode:
          type: string
        campaignId:
          type: string
        campaignName:
          type: ["string", "null"]
        status:
          type: string
          enum: [CREATED, ASSIGNED, OPTIMIZING, AB_TESTING, OPTIMIZED, PUBLISHED, ARCHIVED]
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
        assignedExpertId:
          type: ["string", "null"]
        assignedExpertName:
          type: ["string", "null"]
        aiScore:
          type: ["number", "null"]
        conversionProbability:
          type: ["number", "null"]
        aiReasoning:
          type: ["string", "null"]
        optimizationNote:
          type: ["string", "null"]
        slaDeadline:
          type: string
        slaBreached:
          type: boolean
        notes:
          type: array
          items:
            $ref: "#/components/schemas/CaseNote"
        createdAt:
          type: string
        updatedAt:
          type: string
    CaseListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OptimizationCase"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CaseUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [CREATED, ASSIGNED, OPTIMIZING, AB_TESTING, OPTIMIZED, PUBLISHED, ARCHIVED]
        priority:
          type: string
          enum: [CRITICAL, HIGH, MEDIUM, LOW]
        segment:
          type: string
          enum: [YUKSEK_DEGER, RISKLI_KAYIP, YENI_ABONE, PASIF, BELIRSIZ]
    CaseAssignInput:
      type: object
      required: [expertId]
      properties:
        expertId:
          type: string
    CaseCompleteInput:
      type: object
      required: [optimizationNote]
      properties:
        optimizationNote:
          type: string
    CaseNoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    CaseNote:
      type: object
      required: [id, caseId, authorId, authorName, content, createdAt]
      properties:
        id:
          type: string
        caseId:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        content:
          type: string
        createdAt:
          type: string

    # ── SUBSCRIBER SCHEMAS ────────────────────────────────────────────────────
...